modify: 로그인시 만들어진 공간이 존재하면 해당 공간으로 이동하는 선택지 제공
