+++ client/src/MainPage/MainPage.tsx
@@ -2,6 +2,7 @@ import React, { Suspense, useState } from "react";
 import { Canvas } from "@react-three/fiber";
 
 import Loading from "./Loading";
+import CreateModal from "./components/CreateModal";
 import Header from "../components/Header";
 import UserInfo from "../components/Header/UserInfo";
 import FullScreenModal from "../components/modal/FullScreenModal";
@@ -10,16 +11,11 @@ import { CheckLoggedIn } from "../hooks/useLoggedIn";
 import userStore from "../store/user.store";
 
 import FloatLayout from "../layouts/FloatLayout";
-import NotionIcon from "../assets/images/notion-icon.png";
 import "./style.scss";
 
 export default function MainPage() {
   const [show, setShow] = useState<boolean>(false);
   const { isLoggedIn } = userStore();
-
-  function notionOauthHandler() {
-    window.location.href = "/auth/login";
-  }
 
   function showModal() {
     setShow(true);
@@ -51,13 +47,7 @@ export default function MainPage() {
         </FloatLayout>
       </Suspense>
       <FullScreenModal show={show} css={{ width: "70%", height: "55%" }} setShow={setShow}>
-        <div className="create-modal">
-          <span className="make-gallery">갤러리 만들기</span>
-          <button type="button" onClick={notionOauthHandler}>
-            <img width={25} height={25} src={NotionIcon} />
-            {isLoggedIn ? <span>페이지 가져오기</span> : <span>Notion Login</span>}
-          </button>
-        </div>
+        <CreateModal />
       </FullScreenModal>
     </>
   );
