+++ server/router/galleryRouter.js
@@ -77,10 +77,10 @@ router.get(
   authMiddleware,
   asyncHandler(async (req, res) => {
     const { id } = req.params;
-    console.log("hey!", id);
 
     const result = await loadLastGallery(id);
-    res.status(200).json(result);
+    console.log(result);
+    res.status(200).json( { gallery: result, userID: id} );
   }),
 );
